+++ src/hornbeam_config.erl
@@ -52,6 +52,7 @@
 %%%
 %%% === Python ===
 %%% - pythonpath: Additional Python paths (default: [".", "examples"])
+%%% - venv: Virtual environment path (default: undefined)
 -module(hornbeam_config).
 
 -behaviour(gen_server).
@@ -132,6 +133,7 @@ defaults() ->
         ssl => false,
         certfile => undefined,
         keyfile => undefined,
+        cacertfile => undefined,
 
         %% Protocol
         worker_class => wsgi,
@@ -157,9 +159,11 @@ defaults() ->
         %% WebSocket
         websocket_timeout => 60000,
         websocket_max_frame_size => 16777216,  % 16MB
+        websocket_compress => false,
 
         %% Python
-        pythonpath => [<<".">>, <<"examples">>]
+        pythonpath => [<<".">>, <<"examples">>],
+        venv => undefined
     }.
 
 %%% ============================================================================
@@ -222,7 +226,7 @@ code_change(_OldVsn, State, _Extra) ->
 load_app_env() ->
     Keys = [
         %% Server
-        bind, ssl, certfile, keyfile,
+        bind, ssl, certfile, keyfile, cacertfile,
         %% Protocol
         worker_class, http_version,
         %% Workers
@@ -232,9 +236,9 @@ load_app_env() ->
         %% ASGI
         root_path, lifespan, lifespan_timeout,
         %% WebSocket
-        websocket_timeout, websocket_max_frame_size,
+        websocket_timeout, websocket_max_frame_size, websocket_compress,
         %% Python
-        pythonpath
+        pythonpath, venv
     ],
     lists:foldl(fun(Key, Acc) ->
         case application:get_env(hornbeam, Key) of